+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-## [4.1.2](https://github.com/deltaDNA/android-sdk/releases/tag/4.1.1) (YYYY-MM-DD)
+## [4.1.2](https://github.com/deltaDNA/android-sdk/releases/tag/4.1.2) (YYYY-MM-DD)
+Fixed event sending on push notifications to match Unity/iOS.  
+Fixed missing push notification events in Unity.
 
 ## [4.1.1](https://github.com/deltaDNA/android-sdk/releases/tag/4.1.1) (2016-05-06)
 Fixed activity restarting when opening notifications on Unity.